+++ src/Client/MainComponents/FooterTabs.fs
@@ -69,8 +69,6 @@ let Main (index: int, tables: ArcTables, model: Model, dispatch: Messages.Msg ->
     let id = $"ReorderMe_{index}_{table.Name}"
     let tabRef = React.useElementRef ()
 
-    React.useEffect ((fun () -> console.log state), [| box state |])
-
     /// https://github.com/nfdi4plants/Swate/issues/925
     let startEdit = fun _ -> { state with IsEditable = true } |> setState